Removed model factory from OCP problem maker
Sobec's Model Maker is not used anymore to create the problem.
Instead the cost function can directly be set using the provided
function in the OCP.
Since creating the actual cost function is one of the major features of
this code, it makes more sense to have everything in the package.